Remove DMA signals from FrameBufferIO
The frame buffer interface doesn't need to contain the DMA signals. The GPU actually doesn't care about DMA, it just needs to assert the done signal when a frame has been completed.

The frame buffer DMA can then start a new transfer when the GPU has completed a frame.

The DMA_WAIT state was also removed from the GPU, as it is unlikely that it is actually needed. If it turns out that we need to prevent a new frame from being rendered by the GPU, then we can restore this logic.