Handle ROM offsets in Main module
It doesn't make sense defining the ROM offsets in the Cave module, as this should be described at the level above.

The Cave module should behave similarly to the original arcade hardware, which relies on physical ROM chips. The fact that these ROMs are mapped to different segments of DDR memory is of no concern to the Cave module.