Replace CPU with FX68K implementation
The TG68 core caused stability issues, which have now been addressed by replacing it with the FX68K core.

However, because the FX68K core requires a 32MHz clock, it meant that the system clock frequency had to be reduced to 96MHz. This may have introduced some slowdowns, which are mostly visible when there are lots of sprites being rendered at once.