Fix sprite engine freezing issue
There was an issue where the tile decoder could cause the sprite engine to freeze.

It appeared to be caused by a bug in the TileDecoder state logic, where it could present valid pixel data when the tile ROM output was invalid.

I'm not sure exactly how the freeze occurred, but if the data from the tile ROM was only partially bursted then the TileDecoder could get into an invalid state.