Handle flipping in LayerProcessor
The LayerProcessor needs to handle horizontal/vertical flipping by inverting the pixel position.

The timing of the latch signals also needed to be adjusted, so that the previous tile is loaded, instead of the next one.

Finally, there were a few tweaks to the offset values when the layer is flipped.