Update to Chisel 7.3.0
Update all Chisel code to the latest Chisel (7.3.0).  This has the major
effect of changing the underlying FIRRTL compiler used to build the
Verilog from the Scala-based FIRRTL Compiler (SFC) [[1]] to the MLIR-based
FIRRTL Compiler (MFC or `firtool`) [[2]].  Additionally, as ChiselTest
[[3]] is archived/unmaintained, this switches all the testing to use the
newer Chisel testing framework, Chiselsim.

A number of modules are sensitive to initialization conditions.  This is
exposed when switching to Chiselsim as it will use `initial` block
randomization by default.  Treadle, the default simulator when using
ChiselTest, would initialize everything to zero.  E.g., the cache doesn't
appear to clear its tag bits on reset.  This manifested with one test
failing because the cache would transition to evict instead of
fill (likely because the tag bits indicated that the line was dirty).
Anytime this occurred it likely indicates bugs in the design.  However,
practically these shouldn't manifest as FPGA flows tend to zero-initialize
everything in the bistream.  (And empirically, the core has been working!)

Because of the switch to ChiselSim, this will now pull in a dependency on
Verilator to run tests.

Other miscellaneous changes include fixing deprecations/removed
functions (like the change around `MuxLookup`), removing width inference
of a few output ports for modules used in tests (which is now illegal on
public modules), and updating the mill build flow to use the latest
version of mill.

Verilog generation is now one-file-per-module under the `quartus/cave/`
directory as opposed to a single file `quartus/cave.v`.

[1]: https://github.com/chipsalliance/firrtl
[2]: https://github.com/llvm/circt
[3]: https://github.com/ucb-bar/chiseltest

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@gmail.com>