[ci] Add Verilator CI dependency
Add Verilator to the CI flow as this is necessary for Chiselsim.
Unfortunately, this needs a Verilator version _newer_ than what is
provided by the Ubuntu 24.04 (`ubuntu-latest`) GitHub runner.  That only
provides 5.020 which has a bug in Verilator throwing a spurious
`WIDTHEXPAND` warning (which is fixed in newer versions).

Grab the latest OSS Tabby CAD suite which has a newer version of
Verilator.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@gmail.com>