[rtl] Add Chisel 7.3.0 Verilog, update .qip
Add the newly-generated Cave Verilog files, removing the old `Cave.v`.
Because it is highly advised to have `firtool` generate Verilog
one-file-per-module, that is done here.  All generated verilog files are
put into the `quartus/rtl/cave/` directory.

Update the `quartus/files.qip` to reference these new files.  It is my
understanding that `.qip` doesn't support a wildcard import.  Hence, this
is a bit brittle in that it lists out all the Verilog files one-by-one.
As long as generated Verilog files are committed to this repo, this will
be a maintenance issue.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@gmail.com>